Don't blindly restart statd.
If it's not running for some reason then start it.
If it is running at least with some NFS clients (we suspect FreeBSD)
restarting statd can get lockd in to a state where it stops servicing requests.

There's a bug in lockd and we are chasing it down, but for the moment
this gets TrueNAS NFS HA usable with FreeBSD clients again.