Add a tool that updates HBA firmware
Note: It takes about 3 seconds per controller to run sasXflash
The architecture of this method of firmware updating is to
check if HBA firmware needs to be updated on every boot.

There's a check in place for a previous failed update, so
an HBA that fails the firmware update won't put the system
into a boot loop.

The current code checks the firmware version of each controller
with sasXflash.  So for a system with three controllers we'll
add ~10 seconds to every boot. (plus the fixed cost of firing this
tool up via the rcNG framework)

Next pass is to refactor this so the firmware version check
uses the sysctls (checking them is basically free)
which expose the firmware versions.

Ticket:	#9575