Fix typo that was causing a command to be run incorrectly.
Ticket: 14772

For the want of a shoe the horse was lost...

This fixes a major issue.  Due to a typo NFS was never being detected as
on, which means the exports file wasn't regenerated on the STANDBY
node as it was transitioning to ACTIVE, which would result in it having
an empty exports file.