[NEW_FEATURE] Allow users to install a prebuilt image of RancherOS under bhyve. (#374)
So, we decided to use RanchereOS as our docker system! I have created a RancherOS image that works with bhyve and uploaded it to Ix CDN, this feature basically will download this image, resize(expand) it and run it using bhyve.

There are few tickets related with this request.

Tickets: #23393, #23357 and #23766.