fix(ha): no reason to not kill fenced on forcetakeover
When HA is setup with multiple pools and there is a passphrased pool we
need to run the carp script once the pool is unlocked to make sure
services are restarted in the order we want.
This may interrupt services for the other pool but currently we have no
better way to do this.

Ticket:	#28934