feat(rc.d): --pidfile flag (#948)
* feat(rc.d): --pidfile flag

Previous behavior would overwrite the pid of the daemon process, thus rc.d would kill the wrong process. 

Middlewared no longer has a daemon mode, and always runs in foreground. To allow daemon(8) to control the pidfile, we have a flag for debug to create a pidfile in the absence of daemon(8)

Ticket: #29185

* Remove unused variable

* We don't want TRACE as production default

William noted this to me