Make BMC firmware version check more robust
The BMC firmware for C2750D4I and C2550D4I mainboards shipped in the
FreeNAS Mini product line had a broken watchdog timer that was fixed in
revision 0.30.  The version check we used to decided how to configure
watchdogd assumed the format of the version string reported by ipmitool,
and only checked a portion of the string.

This change performs more checks on the version string, failing more
gracefully if the format is for any reason unacceptable.

Ticket: #35539