Improve validation for jail creation
This commit improves validation when creating jails by ensuring that if a jail with the same name exists, we shouldn't go and try to create another one. Secondly, in case of any failure with jail creation, a call is made to delete any resources which still might be allocated.
Ticket: #41694