tkt-37990: Bug fix for Menu link in Legacy UI (#1743)
* Bug fix for Menu link in Legacy UI

This commit fixes a bug where context was being taken into account for a menu link under System ( View Enclosure ). The value of the tab should've been 'storage.ViewEnclosure', however as it was placed under the System tree, it's value was 'system.ViewEnclosure'. This commit does a context free comparision for children of settings tree.

Ticket: #37990