Bug fix for iscsi extents
This commit fixes a bug where we were not able to update the value of an iscsi extent disk. Secondly it normalizes the output of iscsi.extent.query with iscsi.extent.create enabling users to see the name of the disk which they selected when creating the disk ( if it was specified while creating ).
Ticket: #55692