Configure System Dataset correctly
This commit introduces changes to system dataset service where we don't allow end users to configure a pool for system datasets if they have a passphrase set or are locked. Also if no pool is provided we updated the next best possible pool ourselves. That behaviour has been changed slightly to only look for pools which do not have a passphrase set and then choose one from the remaining pools.
Ticket: #71971