Disable watchdog timer
This commit makes sure that we disable watchdog timer when a shutdown is initiated. Motivation for this step is that we enforce a hard time out limit via sysctl kern.init_shutdown_timeout which is dynamically set by middlewared taking in factor multiple situations i.e no of scripts and their timeouts which will be executed on shutdown.