Reduce default indirect block size from 128K to 32K.
It dramatically reduces write inflation and metadata (de)compression time
in case of random writes.  It may reduce delete/UNMAP performance due
to more head seeks to read the smaller fragmented indirect blocks,
but that operation is prefetchable, so it should not be too bad.

This value originally was set to 16K, but increased to 128K exactly to
improve delete/UNMAP performance.  But that was done before compressed
ARC support, which introduced new factors.

Ticket: #32736