fix(jail/fstab): Add more validation
If an index is not supplied and the action isn’t LIST, we expect a source and destination. In addition we supply a default for source and destination if an index is supplied.

NAS-101531

Signed-off-by: Brandon Schneider <brandon@ixsystems.com>
(cherry picked from commit ab9cc6d3d4d3d2f12e3957d412db1ebdd173090e)