Update list schema to correctly validate/clean items
This commit updates List schema to correctly validate/clean items when the length provided for items is more then one. The motivation behind the changes is that we should validate the given values in the list with all of the schemas provided in items and if we have a positive match for any one of it, we should allow that value in the list.