Allow getting disktemp for non-fatal smartctl exit statuses
Fixes a problem where Freenas report disk temperature "N/A" for one of my disks. The disk in question has smart errors which it got from a failed sata port but is otherwise fully functioning. smartctl reports an exit status of 64 for the disk and therefore doesn't fetch the disk temperature.

https://www.smartmontools.org/browser/trunk/smartmontools/smartctl.8.in#lbAH