Ability to specify jail name when fetching plugin
This commit introduces the ability to specify a custom jail name which can be used to create a plugin. It does not break the existing API and introduces another field for the custom field name which can be passed over. If that is not passed, default behaviour is to use the plugin's name.