Correctly join multiple certificates
This commit fixes a bug where we did not introduce a line break between two certs in a cert chain list. This behavior was okay when almost all certs include a line break at the end but if a cert does not have a line break at end, this fails and nginx is unable to parse the certs correctly.