Introduce Element helper method
This commit introduces a create element helper method which is a wrapper around lxml.etree.Element and allows us to pass text/tail directly and also set children directly. This commit also introduces a new method in vmsupervisor class which will enable us to retrieve a domain's xml through which we can start the domain in question.