Installer cleanup
Remove the ability to upgrade from legacy versions of FreeNAS (9.2 or
earlier). The suggested upgrade path has long been to upgrade to each
consecutive major release. This upgrade process will have repartitioned
the legacy disk format to create the new ZFS-based system, so it is time
we shed the extra complexity.

When importing the freenas-boot pool to check a disk for a an existing
FreeNAS install, import the pool by its GUID taken from the partition's
label instead of by name in case there are multiple pools with the name
"freenas-boot". Likewise, use the GUID when importing the pool to create
a boot-environment during in-place upgrades.

Fix mistaken use of `local readonly` which only creates a local variable
named readonly instead of the intended readonly local variable.

Garbage collect dead code and trailing whitespace.