Add script to resize arbitraty SATA/SAS HDD/SSD.
It resizes specified disk to size in sectors or giga-/terabytes as measured
by HDD/SSD vendors.  If size is not specified, it resizes disk to full size.

SSDs are always wiped in process, HDDs wiped if they support crypto erase.

To be safe script first makes sure that device is not currently open.

Ticket:	NAS-102347