NAS-102573 / 12.0 / Add FSRVP protocol support to middleware (#4020)
Add FSRVP protocol support to middleware

Have middleware configure correct settings for FSRVP protocol support
in Samba. This relies on the vfs module zfs_fsrvp, and allows clients
to manage snapshots on the TrueNAS server through the FSRVP protocol.
Linux and BSD clients can use rpcclient to administer snapshots and
Windows clients can use DiskShadow.exe. The protocol requires clients
be able to generate shares to snapshots on the fly, and so a writable
libsmbconf backend was required for shares. Utilize existing registry
share capabilities in samba to provide this feature.