Remove vfs.zfs.top_maxinflight sysctl.
This sysctl lost its meaning several years ago, doing nothing.  The
real per-vdev limit is vfs.zfs.vdev.max_active, which is set to 1000
by default, and which we may reduce if we want, but that should not
be needed with default settings of other tunables.

It was never set for other than Z50, so we just don't care about it.