Remove asigra plugin
This commit removes asigra plugin from 12 code base as current plugin only performed a migration from asigra service to asigra plugin. Now this migration we expect to take place when the user upgrades from 11.2 to 11.3. That's the official support we are going to give wrt upgrade cycle 11.2->11.3->12. This means that from 11.3->12, users would already have asigra plugin and no migration would be required.