Add initial POSIX1e ACL support to middleware (#4880)
Initial work to read and set a POSIX1e ACL on
filesystems that support that acl type rather than NFSv4 (glusterfs).

More work is needed to add support for maintining Samba ACLs on
POSIX1e filesystems since these ACLs may be written to xattrs in
the security namespace on Linux.

Add acl and cifs-utils packages as dependencies. Former provides
getfacl and setfacl.

Expose acltype property and default to "posixacl" for now (as
opposed to "noacl"). In the future when we add NFSv4 ACL
support to Linux, this will change.