AD: query time from regular DC if NTP broken on PDC emulator
PDC emulator should be authoritative for the domain for domain members,
but if for some reason the NTP server is not working on it, or if it
is unreachable, query time from a DC on the network. If this also
fails, then allow exception to be raised because it's a notable
error condition. Otherwise, only generate warnings in the middleware
log.