AD: cycle through more DCs as part of NTP check
Due to design differences in many Linux AD domains, we should test more
DCs trying to find a working NTP server. The key thing that we're
concerned about here is that our clock hasn't drifted too far from
the kerberos servers in the domains, and provide an error message to
users if it is getting close to the point where kerberos is likely to
break.