Modify migration to simplified SMB configuration setup
Originally in case of non-standard VFS objects we defaulted ACL
to False. This can cause user confusion post-migration if they were
using vfs_zfsacl instead of vfs_ixnas for some reason. Revised
migration strategy is to default to True and only change to False
if vfs_noacl is enabled on the share.