Symlink /usr/share/skel to /etc/skel in FreeBSD
pam_mkhomedir appears to default to trying to get files "/etc/skel".
This possibly indicates a regression in the port, but since it can
be hard to predict behavior of applications that use pam libraries
and hardcode skel path to Linux default location, make this path a
symlink to the FreeBSD path.