Allow keytab setup through NFS plugin
Kerberized NFS requires that the system keytab contain an NFS
Service Principal Name (SPN) entry. For various reasons a user may
need to configure kerberos authentication for NFS at a later date
after initial directory services setup. This PR introduces a method
to do this via the NFS plugin. Currently, only the Active Directory
(AD) directory service is supported.

The API call uses the user-provided credentials to kinit, and then
uses the resulting kerberos ticket for a privileged account to modify
the AD computer account and generate NFS SPN entries for the TrueNAS
server in the AD environment. Once this is complete, the system keytab
/etc/krb5.keytab is updated and the privileged kerberos ticket is destroyed.