Manually toggle value of "passdb backend" in middleware loadparm ctx
It appears that LoadParmContext.load(<smb.conf>) is not sufficient
to catch the second change to Samba's passdb backend (first one is
caught and properly reflected). Fixing this may be a bit hairy and
issue does not actually impact file sharing over SMB (or directory
services). The issue does impact logic internal in middleware for
whether or not to manipulate the passdb and group_mapping databases.
For now, workaround will be to add a new method to manually set this
value as needed in the ldap plugin (and at some future time fix
pyparam).

(cherry picked from commit 26c045acc0788325ac6db0edded24179a8d7b31e)