Add error handling during passdb synchronization
pdbedit can fail to remove entries in the passdb.tdb file if the file
contains garbage data. This creates a situation where only recovery
scenario is the delete the passdb and re-initialize it. Presumably this
error scenario is most likely to occur during entry deletion (not insertion)
and so our error handling will be added to that aspect of passdb handling.
Store copy of garbage passdb.tdb for further analysis if this happens.
Since the NT hashes stored in the passdb.tdb file contain password-equivalent
hashes for local users, keep backup in samba's private dir.