Allow running inadyn in daemon mode
Inadyn init script loads environment variables from /etc/default/inadyn before attempting to start the service and if the env variable dictates not to start it as a daemon, it does not do that. There is no way of overriding this default except for changing it in the default file itself. Correct fix would be to have this fixed upstream and we have an upstream debian bug for this as well ( https://bugs.debian.org/cgi-bin/bugreport.cgi\?bug\=813347 )