Add metadata on chart release operation
This commit adds changes to add metadata to each chart release operation so each chart while rendering k8s resources has access to this metadata. This is useful for cases like upgrade should be performed ( for helm upgrade equals to value changes and upgrade of chart itself which is not true for us ) or specifying storage class which the chart release can consume which offers rollbacks.