use correct logrorate size
https://linux.die.net/man/8/logrotate
Log files are rotated only if they grow bigger then size bytes. If size
is followed by k, the size is assumed to be in kilobytes. If the M is
used, the size is in megabytes, and if G is used, the size is in
gigabytes. So size 100, size 100k, size 100M and size 100Gare all
valid.