Increase crash kernel memory size
With our custom kernel, kdump is not working and in case of a kernel crash, the crash kernel runs out of memory. We have had a user running into this and i was easily able to reproduce as well. When the crash kernel runs out of memory, system does not reboot and gets stuck so it needs to be manually restarted. We should however investigate how we can bring the memory size down, maybe looking at some configuration knobs for the kernel.