Ensure that domain sid component of groupmap entries updated properly (#6778)
Wen Samba's netbiosname changes, the server's local SID is automatically
regenerated and reflected in Samba's passdb. Groupmap entries are not
updated in a similar way. This commit simplifies and fixes smb.fix_sid
so that the domain component of each group's SID matches our current
system SID. If it doesn't, then the group_mapping.tdb is removed and
regenerated. The system SID stored in the TrueNAS configuration database
is also updated at this time. A backgrounded groupmap synchronization
job is started when our netbios name is changed in an SMB service update.
This commit also forces smb4.conf [global] section to be regenerated on
every configuration update (even if SMB is not running). The config file
is read during passdb and groupmap generation and must be kept up-to-date.

(cherry picked from commit df4ed4add07b73614c1042cc3b21dd7252238f9c)