Only raise catalog not healthy alert for preferred trains
This commit adds changes to only raise an alert that a catalog is not healthy if the affected train is set as a preferred train by the user. Motivation for this change is that a catalog might be maintained by lots of devs and each person can be working on their own train for purposes like development, now in the UI users only interact with preferred trains, so it makes sense that we only let the user know that a catalog is not healthy if one of the affected trains is what the user has set his preferred train.