Increase timeout to start k3s service
This commit adds changes to increase timeout to start k3s service. First time k3s is initialized it takes some time before it's systemd unit actually reports that it's active and 20secs was good enough - however Kris was running into issues on a 4gb ram / 1 core vm and there the timeout of 40 secs works. This is less then ideal for waiting as ideally the systemd unit file should not be doing this but that's an upstream issue ( will be creating an issue about it upstream ).