Remove dependency on samba3 loadparm context (#7351)
Initializing a samba3 loadparm context requires that middleware
open registry.tdb and initialize a full s3 context. This PR
switches us to use a different samba python module that is
lighter-weight and doesn't require the tdb open. As a side-effect
it's utility is more limited in scope. It can be used to
look up default values for parameters (which may change between
samba releases), and it can be used to validate SMB auxiliary
parameters. The key limitation to it is that the limited
loadparm context can't be used to read parameters that are stored
in the static configuration file (/etc/smb4.conf) and so we
need an alternate method to read non-default parameters in that
file. In this case, we simply read and convert into JSON types.
There should only be a limited number of well-known parameters
inserted here (unless users are manually editing the files) and
so this doesn't need to be perfect at typing all the different
parameters in the smb4.conf. If the requested parameter isn't
in the smb4.conf file, then we use loadparm context to look up
the default value and return that.