Skip ALLOW test for NFS4 WRITE_ATTRIBUTES on SCALE (#7402)
At this point, special handling of this permission is not
ipmlmented. Specifically, WRITE_DATA | WRITE_ATTRIBUTES are
required in order to alter timestamps. At this point no
adverse affect has been observed with SMB server and such
behavior is comparable with FreeBSD NFSv4 ACL implmenetation.
We can re-evaluate in future if it becomes a demonstrable
issue in real-life workloads, otherwise we'll keep
implementation simpler.