NAS-112847 / 22.02-RC.2 / Fix invalid realm info in auto-generated realm and add job progress to activedirectory.leave (#7676)
kerberos.realm.direct_create bypasses the plugin's compress
method and so we need to insert empty strings for kdc, admin server,
and kpasswd server.

The activedirectory.leave method was converted into a middleware job,
but progress messages were not added. Add these and also ensure
that the idmap service is restarted after the domain is left and Samba
reconfigured.