NAS-113054 / 22.02-RC.2 / Add basic auto-configuration for non-standard LDAP servers (#7762)
Detect from root DSE whether the LDAP domain is AD or FreeIPA. If we're
in AD domain, log a warning, but proceed with some auto-configuration
to provide minimal parameters so that NSS works. This provides
framework for validating LDAP + kerberos in various forms against
an AD DC in the CI pipeline.

The PR also fixes a couple of bugs in LDAP + kerberos that were
discovered in testing.