Expand `__[POSTCOMPILE_encrypted_provider_1` into valid SQL syntax
SQLAlchemy includes a variant on a bound parameter known as `BindParameter.expanding`,
which is a “late evaluated” parameter that is rendered in an intermediary state when
a SQL construct is compiled, which is then further processed at statement execution
time when the actual known values are passed. “Expanding” parameters are used for
`ColumnOperators.in_()` expressions by default so that the SQL string can be safely
cached independently of the actual lists of values being passed to a particular
invocation of `ColumnOperators.in_()`.

We need to render the `IN` clause with real bound parameter symbols as we store these
statements in HA database journal and execute them in a lower-level connection that
does not support these SQLAlchemy substitutions.