Prevent contents of /etc/exports.d from altering config
knfsd / exportfs will attempt to load any exports file added
to /etc/exports.d. This can lead to filesystems being
unexpectedly exported and a broken NFS configuration.

In this PR, we add an additional check when generating
exports where we remove any unexpected file from
/etc/exports.d and then make the directory immutable.

If for some reason, a file cannot be removed, then we
intentionally generate an invalid exports config with
a comment stating that the user needs to clean up whatever
he or she did there.

When a zpool is imported we force-inherit mountpoint property.
This has the effect of trying to sharenfs if user has set
the sharenfs dataset property. We now have the immutable flag
set on /etc/exports.d. This causes zfs_prop_inherit to
fail with error SHARENFSFAILED (although mountpoint does get
set correctly -- prop is inherited before zfs_sharenfs() is
called).

This PR adds special handling for when mountpoint property
is being inherited. If call fails with error code of
SHARENFSFAILED or SHARESMBFAILED, we set the CallError
errno to a special value that is then used to tell
pool.import_pool that shares need to be disabled.

(cherry picked from commit db5e458dd223956c15f145312021a98d0a2b2a04)