Optimize acl_is_trivial
This method is called when we call filesystem.listdir and
filesystem.stat. Unfortunately because of how it was implemented
for cases with relatively largish numbers of files, it would
become quite slow. filesystem.listdir took ~30 seconds for a
directory with 1000 files. Now that the ZFS xattr handler has
been updated for listdir to include the synthetic system.nfs4_acl_xdr
xattr in the return buffer if the ZFS ACL is non-trivial,
we can directly check this via os.listxattr().

With this new optimization, filesystem.listdir on the same directory
now takes less than 1 second.

(cherry picked from commit 5ce33ca022acf575cac465dc54371278faf5c362)