Add new helper function to convert mountinfo to dict (#9200)
We were using libzfs to convert paths to dataset name. Internally
libzfs does this thorugh stat() on the path and looking up
mount entries. We can instead do this ourselves without having
to use a libzfs handle.

To facilitate this add a new general-purpose method to return
a dict containing mount information indexed by devid. The
reason to index this way is because it allows efficient lookup
of an entry based on stat / lstat output. Refactor
filesystem.statvfs to use this as well.