reporting: Remove dummy memory-cache statistic
The memory-cache statistic reported the size of the virtual memory page
cache queue.  The page cache queue has been removed in favor of the
page laundry queue.  We added a stat for this but failed to remove the
page cache stat, which is now only a dummy stat left for compat and
always reports 0.

Remove the dummy memory-cache statistic.