Fix initialization event for ctdbd (#9616)
For various reasons the middleward socket may not be available
when we are starting the ctdbd service. If this is the case, then
we should fall back to just checking that the ctdb_shared_volume
gluster volume is mounted.

Additionally, this PR makes it so that we avoid initializing a
middleware client connection during event processing if the
connection will not be needed.