Improve wireguard retrying logic
This commit adds changes to improve wireguard retrying logic. Currently what we did was that when truecommand service was started, wireguard health check was triggered right after starting the service which in the unlikely event TC container was down, would fail and poll the TC API to see if the API key is active - which in this case would return as active and try to start TC service again which will result in the same vicious loop. What we do now instead is, that we wait a pre-defined time before checking wireguard connection health giving some grace period to TC container to come up and also not continously perform a wireguard start loop.