Improve kerberos tests (#10338)
This PR updates the regression tests for
kerberos when joined to AD. We use new active_directory
context manager to ensure proper test cleanup.

Tests are now expanded to included ticket management.
After joining AD we re-kinit with a short-lived kerberos
ticket and then verify that `kerberos.renew` renews the
ticket properly. The tests also validate that our
background job to monitor pending kerberos tickets is
reporting the correct information.

During the course of testing, an edge-case was encountered
where sysvol replication for the fresh AD account had not
ocurred by the time we tried to get a limited kerberos ticket.
As a result of this, kerberos.do_kinit was expanded to allow
for overriding / hard-coding a particular KDC in the krb5.conf.
This is used to temporarily lock us to a single KDC (the one
we used to join AD) that that we can reliably kinit. This
temporary stub krb5.conf is then cleared when we generate
a full krb5.conf file after downgrading our ticket.

Co-authored-by: Andrew Walker <awalker@ixsystems.com>