Flush samba gencache after removing user or group (#10651)
If group is removed from file then NSS query will
fall through to nss_winbind, which may have cached
result. In principle, winbind shouldn't be giving
a response here, but fix will be more involved and
so for now we will simply flush cached entries on
user or group removal.