Add migration for hostname_virtual to netbiosname (#10876)
In legacy TrueNAS appliances we supported having separate netbiosnames
for each controller and joining each storage controller separately to
Active Directory. Unfortunately, this legacy configuration had a
tendency to break client assumptions about session and server
persistence during failover events, which necessitated a redesign to
use a single netbiosname for both controllers with Samba's private
state files migrating between storage controllers during a failover
event.

The legacy configuration using separate netbiosnames was deprecated in
TrueNAS 11.3 and support for it finally removed in its entirety in
SCALE BlueFin. Due to how backwards-compatibility was implemented for
legacy configurations in 11.3 and later, the hostname_virtual field was
used as a substitute for the netbiosname when generating Samba
configuration.

SCALE BlueFin contains a database migration that removed the legacy
netbiosname_b field from the SMB configuration and so the logic to
overlay hostname_virtual field was removed.

Due to this change though we need to overwrite the netbiosname with the
hostname_virtual if it exists via a database migration. This ensures
that AD users upgrading to SCALE from TrueNAS 13 Enterprise will not
lose AD integration.

The migration happens if the following two conditions are met:

1) hostname_virtual is populated - this only happens in HA platform
2) hostname_virtual is a valid netbiosname