NAS-122782 / 23.10 / Remove hard-coded ctdb_shared_vol for glusterfs (#11615)
Ctdb and clustered middleware configuration relies on a small numbers of
configuration files that are stored within a gluster volume. Originally,
we used a hard-coded dedicated volume. This PR makes the location for
these files configurable and provides a framework to shift volume
location. This is a significantly disruptive operation cluster-wide and
should only be performed in a maintenance window. Default volume will
be first volume created.