Properly catch exception with invalid helm secret
We have seen cases earlier with users where helm had malformed secrets which resulted in us not being able to parse those secrets. Currently we have seen a user where another exception was being raised and this commit handles that exception so we properly handle this. This change makes sure that other apps which don't have a malformed secret are not affected by this inconsistency on helm side.