NAS-122813 / 23.10 / Add more protection for JWT usage in cluster event (#11622) (#11626)
Initial implementation of JWT usage was to encode static dummy payload
with pre-shared key. Presence of decodable header was used to
authenticate the command / payload sent to the webhook. Since the
encoded payload and secret were both static, the JWTOKEN used for all
commands was identical.

This PR replaces the static dummy payload with a dict containing
timestamp and uuid for message id. If clock slew from message we receive
is more than 10 seconds or if the message id is reused, then we reject
the event request.

Co-authored-by: Andrew Walker <awalker@ixsystems.com>