NAS-123585 / 22.12.4 / Add validation for share ACL payload (#11983)
This add checks that SID, domain, and name is known to us on an
as-needed basis. Checking early in middleware gives better control
of error messages before we pass along to the sharesec command.
This endpoint has been substantially rewritten for Cobia and so
the change only applies to BlueFin.