Safely validate unlocked datasets
This commit adds changes to not raise a validation error if a dataset is specified explicitly in the list which is already unlocked. We are not erroring out here because it should be fine as rest of the implementation is not concerned about it being unlocked already and a consumer to be on the safe side can just specify whichever datasets he wants unlocked rather then checking first and only specifying locked ones.