Fix create_ancestors when share_type is SMB (#12393)
This commit modifies our execute check so that we by default skip
non-existent path components. This is relevant when validating case
where ZFS create_ancestors is specified when creating a new dataset. We
don't want to fail validation on the non-existent path component, but
we do want to validate permissions on any parent datasets that do exist.