Convert check for presence of DS_TYPE_DEFAULT_DOMAIN (#12512) (#12514)
Initial check in this migration was for primary key == 5. This
has been hard-coded since FreeNAS 9.3, but in an early angelfish
beta a code change was introduced that automatically deleted
table entries to prevent update failures in some edge cases. This
resulted in a required default idmap table entry being deleted for
some users (the precise details are nuanced and covered in the
related commits / jira tickets). The situation was rectified and
a migration introduced to automatically re-add the missing table
entry. It appears that in some cases, users manually added the
entry back via shell, but didn't preserve the original primary
key that was used to identify the special entry. This has
resulted in a handful of users that are unable to upgrade.

Co-authored-by: Andrew Walker <awalker@ixsystems.com>