Fix k8s pod stats integration test (#12898)
This commit fixes 2 issues which we were seeing with k8s pod stats test. First issue was that sometimes netdata has not yet collected any stats from Kubernetes, and the test is attempting to retrieve non-existent stats, resulting in an error. The second problem is that when netdata does get stats, it initially provides zero values on the first update, leading to assertion errors in the test.

Here what we do now instead is that we make sure the cluster is up before proceeding with asserting stats and also ensure that the stats have been updated twice to make sure it has actually retrieved data which is relevant.