Fix permissions check for services plugin using api key (#13213)
Allow API keys to toggle services. We only need to check whether
this is a user session since API keys don't support roles. This
means that once an API key is within the permissions check function
we already know that it has been explicitly granted access to this
endpoint via its allowlist.