Fix temporary mount of system dataset during move (#13286)
When moving the system dataset to a new pool we first temporarily
mount the system dataset in the middleware run directory before
running rsync to copy all of the data to the new path. During
a recent refactor of the system dataset heirarchy the nuance of
the temporary mount point was lost.