Loosen restrictions for LDAP idmap validation (#13638)
When generating synthetic accounts (ones used for user.query and
group.query results) stop requiring the LDAP UID and GID to fit within
the allocated idmap ranges in winbindd. We can do this safely now that
we return whether group is `local` in nss module results. This is first
step toward removal of legacy samba NT domain support as well.