NAS-128003 / 24.10 / Update zvol collection method (#14076)
Updated the way we collect the zvols.  The previous method did not do a depth search.
Fixed a few typos in TRUENAS-MIB.txt which includes regenerating TRUENAS-MIB.py.
Added a context manager for simple 'file' creations in the filesystem asset.
Created a CI test in test_440_snmp.py for this issue.