Improve auditing for middleware methods (#14158)
Populate `username` and `address` fields in middleware audit
entries based on information from the middleware session.

The UI prominently displays the `username` field which is currently
hard-coded to root. This commit adds a helper function to generate
a username string based on the current authenticated credentials
for the session generating the audit entry.