Handle potential case when tos can be null (#14484)
This commit fixes a case where tos can be null which we get from an ACME provider. In the database we keep it as an empty string because this needs to go through the libraries we are using to generate an ACME certificate which can/cannot error out and unfortunately we don't have an ACME provider atm which gives a null tos to test a null case. Keeping it as an empty string in that case in database if tos received is null.