Fix recursion in getting SID info (#14514)
Early in electric eel development, the SID values associated with
local users and groups were changed to be deterministic based on
their respective datastore table primary keys. A subsequent commit
changed how this SID information was retrieved in user.get_user_obj
and group.get_group_obj to ensure that we stopped relying on
a winbind client connection (to robustize against edge case of
user not having winbindd running). This SID information was obtained
via user.query and group.query.

Unfortunately, the developer neglected to set an additional
['local', '=', True] filter for the query request (which limits to
local users only), which resulted in call to
directoryservices.cache.query when AD or LDAP is enabled. If the
account is a local account then the cache lookup failure results in call
back into user.get_user_obj or group.get_group_obj resulting in loop.

This commit correctly limits the lookup to local users and groups.