NAS-131456 / Update user-group memberships using datastore (#14601)
It seems that the original code for updating user-group relationships (that was never touched since
it was written in 2017) contains an error:

https://github.com/truenas/middleware/blob/master/src/middlewared/middlewared/plugins/account.py#L1367

gms is an array of account.bsdgroupmembership objects, so gm["id"] has a Row ID type for
account.bsdgroupmembership table. groups is an array of Row IDs for account.bsdgroup table. gm['id']
not in groups check has no sense, it compares Row IDs of different tables.

Our datastore plugin has the ability to manage many-to-many relationships table, let's use it.