Prevent id-assignment foot-shooting in accounts (#14746)
After internal discussion the valid use-cases for allowing groups
with duplicate gids are few and far between, but the risk of novice
users accidentally exposing themselves to unexpected disclosure of
information is not obscure. Hence, we are removing this quasi-feature
from the v25.04 API before it's finalized. This same argument
for allowing API consumers to change the uid for users and gid for
groups can be made. In this case the functionality was never exposed
via the webui.