Handle case where netbiosalias is omitted from AD payload (#15013)
When netbiosalias is set in SMB configuration and omitted from the
payload in the AD form, it gets populated as an empty list causing
a ValidationError to be raised (because we don't allow changing
netbios configuration while AD is enabled). Unfortunately, some
howto guides online have recommended users to set the netbiosname
netbios alias to the same value, and as is the way of such things
this advice is both unnecessary and detremental. This means that users
who followed such guides are now unable to modify their AD configuration
through the webui (which no longer sets the netbiosalias by default).

Set the default for netbiosalias to None so that it does not default
to an empty list as a special value indicating that the existing
netbios aliases should be preserved.

(cherry picked from commit 4226eb0d22aef2358bfdc0c59d1adc23e9c355a6)

Co-authored-by: Andrew Walker <awalker@ixsystems.com>